Catalog complete; untested yet

+++ backend/src/exemplary/exemplary-router.js
@@ -10,7 +10,7 @@ module.exports = function(app){
         const result = new ExemplaryModel().get_all(current_page, page_size)
             .then(function(result) {
                 let form_result = {
-                    "count": result[0].full_count,
+                    "count": result[0] === undefined ? 0 : result[0].full_count,
                     "results": result
                 }
                 return res.json(form_result)
@@ -18,6 +18,23 @@ module.exports = function(app){
                 console.log(err)
             })
     })  
+
+    app.get('/api/exemplary/full', async (req, res, next) => {
+
+        const current_page = parseInt(req.query.currentPage)
+        const page_size = parseInt(req.query.pageSize)
+        
+        const result = new ExemplaryModel().get_exemplaries(current_page, page_size)
+            .then(function(result) {
+                let form_result = {
+                    "count": result[0] === undefined ? 0 : result[0].full_count,
+                    "results": result
+                }
+                return res.json(form_result)
+            }).catch(function(err) {
+                console.log(err)
+            })
+    })
 
     app.get('/api/exemplary/author/:id', async (req, res, next) => {
         const id = parseInt(req.params.id)
@@ -27,7 +44,6 @@ module.exports = function(app){
                 return res.json(result)
             }).catch(function(err) {
                 console.log(err)
-                return next(e)
             })
     })
 
@@ -39,7 +55,6 @@ module.exports = function(app){
                 return res.json(result)
             }).catch(function(err) {
                 console.log(err)
-                return next(e)
             })
     })
 
@@ -47,9 +62,7 @@ module.exports = function(app){
 
         const input_data = req.body;
 
-        if (!input_data) {
-            return res.status(400).end();
-        } else {
+        if (input_data !== undefined)  {
                 
             const result = new ExemplaryModel().insert_exemplary(input_data.author, input_data.book)
             .then(function(result) {
@@ -57,9 +70,30 @@ module.exports = function(app){
                 return res.json(result)
             }).catch(function(err) {
                 console.log(err)
-                return next(e)
             })
 
+        } else {
+            return res.status(400).end()
+        }
+
+    })
+
+    app.post('/api/exemplary/b/', async (req, res, next) => {
+
+        const input_data = req.body;
+
+        if (input_data.data !== undefined)  {
+                
+            const result = new ExemplaryModel().insert_exemplary_bulk(input_data.data)
+            .then(function(result) {
+                console.log("added exemplary bulk")
+                return res.json('sucess')
+            }).catch(function(err) {
+                console.log(input_data.data)
+            })
+
+        } else {
+            return res.status(400).end()
         }
 
     })
@@ -68,18 +102,17 @@ module.exports = function(app){
 
         const input_data = req.body
 
-        if (!input_data) {
-            return res.status(400).end();
-        } else {
+        if ((input_data.new_data !== undefined) && (input_data.old_data !== undefined)) {
                 
             const result = new ExemplaryModel().update_exemplary(input_data.new_data, input_data.old_data)
             .then(function(result) {
                 return res.json('updated author row')
             }).catch(function(err) {
                 console.log(err)
-                return next(err)
             })
             
+        } else {
+            return res.status(400).end()
         }
 
     })
@@ -88,18 +121,36 @@ module.exports = function(app){
 
         const input_data = req.body
 
-        if (!input_data) {
-            return res.status(400).end();
-        } else {
+        if ((input_data.author !== undefined) && (input_data.book !== undefined)) {
                 
             const result = new ExemplaryModel().delete_exemplary(input_data.author, input_data.book)
             .then(function(result) {
                 return res.json('deleted author row')
             }).catch(function(err) {
                 console.log(err)
-                return next(err)
             })
             
+        } else {
+            return res.status(400).end()
+        }
+
+    })
+
+    app.post('/api/exemplary/rem/b', async (req, res, next) => {
+
+        const input_data = req.body
+
+        if (input_data.data !== undefined)  {
+                
+            const result = new ExemplaryModel().delete_exemplary_bulk(input_data.data)
+            .then(function(result) {
+                return res.json('deleted author rows')
+            }).catch(function(err) {
+                console.log(err)
+            })
+            
+        } else {
+            return res.status(400).end()
         }
 
     })
